+++ Kernel/Modules/AgentTicketCompose.pm
@@ -1685,7 +1685,20 @@ sub _Mask {
 
         my %Selected;
         if ( $Param{ArticleTypeID} ) {
-            $Selected{SelectedID} = $Param{ArticleTypeID};
+
+            # use email-internal article type if replying to article with internal type; otherwise use email-internal
+            my $ArticleTypeSelected = $TicketObject->ArticleTypeLookup( ArticleTypeID => $Param{ArticleTypeID} );
+            if ( $ArticleTypeSelected ) {
+                if ( $ArticleTypeSelected =~ m{internal} ) {
+                    $Selected{SelectedID} = $TicketObject->ArticleTypeLookup( ArticleType => 'email-internal' );
+                }
+                else {
+                    $Selected{SelectedID} = $TicketObject->ArticleTypeLookup( ArticleType => 'email-external' );
+                }
+            }
+            else {
+                $Selected{SelectedValue} = $Config->{DefaultArticleType};
+            }
         }
         else {
             $Selected{SelectedValue} = $Config->{DefaultArticleType};